fix: next wave of fixing tests

+++ vyper/simple/algorithm/arrays/multidimensional.vy
@@ -46,13 +46,13 @@ def array_2d() -> uint256:
     array: uint8[ARRAY_SIZE_2D][ARRAY_SIZE_2D] = empty(uint8[ARRAY_SIZE_2D][ARRAY_SIZE_2D])
 
     for i: uint8 in range(0, ARRAY_SIZE_2D):
-        for j: uint256 in range(0, ARRAY_SIZE_2D):
+        for j: uint8 in range(0, ARRAY_SIZE_2D):
             array[i][j] = i*ARRAY_SIZE_2D + j
 
     result: bool = True
     # other order
 
-    for j: uint256 in range(0, ARRAY_SIZE_2D):
+    for j: uint8 in range(0, ARRAY_SIZE_2D):
         for i: uint8 in range(0, ARRAY_SIZE_2D):
             result = result and array[i][j] == i*ARRAY_SIZE_2D + j
 
@@ -67,15 +67,15 @@ def array_3d() -> uint256:
     array: uint8[ARRAY_SIZE_3D][ARRAY_SIZE_3D][ARRAY_SIZE_3D] = empty(uint8[ARRAY_SIZE_3D][ARRAY_SIZE_3D][ARRAY_SIZE_3D])
 
     for i: uint8 in range(0, ARRAY_SIZE_3D):
-        for j: uint256 in range(0, ARRAY_SIZE_3D):
-            for p: uint256 in range(0, ARRAY_SIZE_3D):
+        for j: uint8 in range(0, ARRAY_SIZE_3D):
+            for p: uint8 in range(0, ARRAY_SIZE_3D):
                 array[i][j][p] = i*ARRAY_SIZE_3D*ARRAY_SIZE_3D + j*ARRAY_SIZE_3D + p
 
     result: bool = True
     # other order
 
-    for p: uint256 in range(0, ARRAY_SIZE_3D):
-        for j: uint256 in range(0, ARRAY_SIZE_3D):
+    for p: uint8 in range(0, ARRAY_SIZE_3D):
+        for j: uint8 in range(0, ARRAY_SIZE_3D):
             for i: uint8 in range(0, ARRAY_SIZE_3D):
                 result = result and array[i][j][p] == i*ARRAY_SIZE_3D*ARRAY_SIZE_3D + j*ARRAY_SIZE_3D + p
 
@@ -90,17 +90,17 @@ def array_4d() -> uint256:
     array: uint8[ARRAY_SIZE_4D][ARRAY_SIZE_4D][ARRAY_SIZE_4D][ARRAY_SIZE_4D] = empty(uint8[ARRAY_SIZE_4D][ARRAY_SIZE_4D][ARRAY_SIZE_4D][ARRAY_SIZE_4D])
 
     for i: uint8 in range(0, ARRAY_SIZE_4D):
-        for j: uint256 in range(0, ARRAY_SIZE_4D):
-            for p: uint256 in range(0, ARRAY_SIZE_4D):
-                for h: uint256 in range(0, ARRAY_SIZE_4D):
+        for j: uint8 in range(0, ARRAY_SIZE_4D):
+            for p: uint8 in range(0, ARRAY_SIZE_4D):
+                for h: uint8 in range(0, ARRAY_SIZE_4D):
                     array[i][j][p][h] = i*ARRAY_SIZE_4D*ARRAY_SIZE_4D*ARRAY_SIZE_4D + j*ARRAY_SIZE_4D*ARRAY_SIZE_4D + p*ARRAY_SIZE_4D + h
 
     result: bool = True
     # other order
 
-    for h: uint256 in range(0, ARRAY_SIZE_4D):
-        for p: uint256 in range(0, ARRAY_SIZE_4D):
-            for j: uint256 in range(0, ARRAY_SIZE_4D):
+    for h: uint8 in range(0, ARRAY_SIZE_4D):
+        for p: uint8 in range(0, ARRAY_SIZE_4D):
+            for j: uint8 in range(0, ARRAY_SIZE_4D):
                 for i: uint8 in range(0, ARRAY_SIZE_4D):
                     result = result and array[i][j][p][h] == i*ARRAY_SIZE_4D*ARRAY_SIZE_4D*ARRAY_SIZE_4D + j*ARRAY_SIZE_4D*ARRAY_SIZE_4D + p*ARRAY_SIZE_4D + h
