+++ vyper/simple/algorithm/sort/merge_iterative.vy
@@ -54,11 +54,11 @@ def mergeSort(_array: uint8[ARRAY_SIZE], direction: uint8) -> uint8[ARRAY_SIZE]:
     size: uint8 = 1
     tmp: uint8[ARRAY_SIZE] = empty(uint8[ARRAY_SIZE])
 
-    for _: uint256 in range(ARRAY_SIZE):
+    for _: uint8 in range(ARRAY_SIZE):
         if not size < ARRAY_SIZE:
             break
         ptr: uint8 = 0
-        for __: uint256 in range(ARRAY_SIZE):
+        for __: uint8 in range(ARRAY_SIZE):
             if not ptr < ARRAY_SIZE:
                 break
             mid: uint8 = ptr + size
@@ -72,7 +72,7 @@ def mergeSort(_array: uint8[ARRAY_SIZE], direction: uint8) -> uint8[ARRAY_SIZE]:
             ptr1: uint8 = ptr
             ptr2: uint8 = ptr + size
 
-            for ___: uint256 in range(ARRAY_SIZE):
+            for ___: uint8 in range(ARRAY_SIZE):
                 if not (ptr1 < mid and ptr2 < last):
                     break
                 if direction == DirectionNone or (direction == DirectionAscending and array[ptr1] < array[ptr2]) or (direction == DirectionDescending and array[ptr1] > array[ptr2]):
@@ -83,14 +83,14 @@ def mergeSort(_array: uint8[ARRAY_SIZE], direction: uint8) -> uint8[ARRAY_SIZE]:
                    ptr2 += 1
                 ptr += 1
 
-            for ___: uint256 in range(ARRAY_SIZE):
+            for ___: uint8 in range(ARRAY_SIZE):
                 if not ptr1 < mid:
                     break
                 tmp[ptr] = array[ptr1]
                 ptr += 1
                 ptr1 += 1
 
-            for ___: uint256 in range(ARRAY_SIZE):
+            for ___: uint8 in range(ARRAY_SIZE):
                 if not ptr2 < last:
                     break
                 tmp[ptr] = array[ptr2]
